Set viper to PR #635, SetKeysCaseSensitive
https://github.com/spf13/viper/pull/635

YAML keys are case sensitive according to the specification, Viper has historically forced lowercase to make lookups more predictable. This becomes a problem when the YAML is then exported again and the consumer is case sensitive.
The Viper PR to which this sets the module to seems to resolve this.